Added more style items
Managed to do a decent header; made a footer, will stay at the bottom of the page. Also added a gradient and drop shadow to the header. Tried to make the main content be in a rounded rectangle -- still doesn't work properly, must check this. Must also check how the header is seen in a dif. screen size :/